Alter state passing between components to be much more react-like
Instead of calling a function in onchange and doing things in there, now passing the keywords into a searchresults component which then does the matching and passes the games that should be shown to the gametable component.
It actually feels much faster, and loading all 4000 ps2 titles into the browse the search is still snappy.