#48 Hide animation checkbox when there is none, allow user to scrub through frames when animation paused
This introduces a new callback function from the model renderer implementation (ModelRendererImpl.tsx) to the icon viewer page (Icon.tsx).
The callback tells the icon viewer page how many frames there are, or 0 if there is no animation.
This callback could have extra data added to it for use in the future.
Also add a small fix to homepage to fix discrepancy between the progress bar percentage and the one displayed below.