Upload icons contributed by IdiOT in the Discord
Skipped Max Payne as it was already present.
Overwrote Vice City as it was using an the old format.
Added comment for Torrente 3 being the Spanish version, need to think about how to model variants like this.
Set 'The Sims' icon count to 1 despite being 3, because they are visually the same (bouncing house icon).